requirements.txt
fastapi
uvicorn
pymilvus
sentence-transformers
torch
openai>=1.0.0
python-dotenv